+++ src/__shared__/ui/shadcn/dialog.tsx
@@ -4,7 +4,7 @@ import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { X } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/__shared__/utils/utils";
 
 const Dialog = DialogPrimitive.Root;
