+++ src/__shared__/ui/shadcn/form.tsx
@@ -10,8 +10,8 @@ import {
   useFormContext,
 } from "react-hook-form";
 
-import { Label } from "@/components/ui/label";
-import { cn } from "@/lib/utils";
+import { Label } from "@/__shared__/ui/shadcn/label";
+import { cn } from "@/__shared__/utils/utils";
 
 const Form = FormProvider;
