+++ src/__shared__/utils/db/index.ts
@@ -1,0 +1,13 @@
+import { PrismaClient } from "@prisma/client";
+
+/**
+ * @public
+ */
+export type TransactionClient = Omit<
+  PrismaClient,
+  '$connect' | '$disconnect' | '$on' | '$transaction' | '$use'
+>;
+/**
+ * @public
+ */
+export const db = new PrismaClient();
