+++ src/__shared__/utils/result/index.ts
