+++ src/app/(operation-check)/operation-check/address/_components/address-form/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { IFindOneAddressResponse } from "@/backend/address/use-case/find-one-address/response-interface";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/__shared__/ui/shadcn/button";
 import {
   Form,
   FormControl,
@@ -10,9 +10,9 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { useToast } from "@/components/ui/use-toast";
+} from "@/__shared__/ui/shadcn/form";
+import { Input } from "@/__shared__/ui/shadcn/input";
+import { useToast } from "@/__shared__/ui/shadcn/use-toast";
 import { env } from "@/env.mjs";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { atom, useAtom } from "jotai";
