+++ src/app/(operation-check)/operation-check/auth/_components/logout-button/index.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/__shared__/ui/shadcn/button";
 import { signOut } from "next-auth/react";
 
 export const LogoutButton = () => {
