+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/lib/auth/auth-options";
+import { authOptions } from "@/__shared__/utils/auth/auth-options";
 import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
