--- src/lib/db/index.ts
@@ -1,6 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-/**
- * @public
- */
-export const db = new PrismaClient();
